login and logout is written for website package.json updated for linux phyton3 systems. Sessions not implemented
