+++ buroute-web/src/login_signup_style.css
@@ -140,6 +140,19 @@ input[type="password"] {
 
 } 
 
+.error-message {
+  color: white; 
+  background-color: rgba(244, 67, 54, 0.8);  
+  border-radius: 4px; 
+  padding: 4px; 
+  margin-top: 10px; 
+  text-align: center;  
+}
+
+.error-message:empty {
+  display: none; 
+}
+
 @media (max-width: 600px) {
   .container_center {
     width: 90%;
