+++ wiki_database_API/database_search/urls.py
@@ -10,5 +10,6 @@ urlpatterns = [
     path('users/<int:pk>/', views.user_detail, name='user_detail'),
     path('create_user/', views.create_user, name='create_user'),
     path('create_node/', views.create_node, name='create_node'),
-    path('login/', views.login_user, name='login_user')
+    path('login/', views.login_user, name='login_user'),
+    path('logout/', views.logout_user, name='logout_user'),
 ]
